Remove Header() from rtcp.Packet
This will allow CompoundPacket to become a Packet.

It's also redundant. The reason we exposed Header was to allow
fetching the type name specified in the Header. Instead, users
can perform a type switch to detect the underlying type. If the
type has not been implemented, you can cast to RawPacket and
fetch the Header there.

Relates to #25