Adding String() method to Packet interface
Nearly all RTCP packet types already include String() methods, which
are quite valuable for diagnostic and logging purposes. This change
updates the base Packet interface to include String() so that it
can be used by applications without having to perform a complex set
of type assertions.

As a minor bit of housekeeping, it also moves assertions that the
Packet interface is satisfied from implementation files into test
files.