Add support for RTCP Extended Reports
This patch adds baseline support for RTCP extended reports, as defined
in RFC 3611, including the baseline seven report block types it
describes. This patch also adds some introspection-based support
functions to provide semi-automated marshaling and unmarshaling of
arbitrary structures, and automated stringifying of Packet structures,
so as to greatly simplify adding the other 27 report types.
This patch addresses the baseline functionality of Issue #8.